update
Former-commit-id: 2364e857e9447ab2386aeb5e6c02f98fe06d157e [formerly 0e942bb6003b6ef594ffa723a64defc1c6230bbf] [formerly cb4d351638ead1df46bd55a619ae58ccf461b58b [formerly 5f878fc7715ffba0061aabc322e7bca4fdbf44d7]] [formerly 6280310a3267d9632ea47d8e095623155c076cf4 [formerly e38218b4bf97596c15cdf5188516cb5d03596d1b] [formerly d803703c51688237d75f8557256a49b16666c56a [formerly 7b5626f0fd5f707587370b6649807864caf15b51]]] [formerly e1004ce2face16c8cdf45adc3ec3a054d391ba6f [formerly 992fd985fdf4b40581603e4081dbb7faf5635364] [formerly af67ad0b3549d5032ea32c9bb9e377578414bbf5 [formerly 1967e97c9bbd7b3ce95f9ac3306df2f0b21e7605]] [formerly 5f1e225f0854b249c3e04f54fb8eddaee9600ded [formerly 939e96baa3c8e5901c21dc72d4252d1a6b9603fd] [formerly a5e97864c800603adc118b8c35c3621ff7ef02f9 [formerly d62763b45a2cf5df39703da7b66bcc43d6e254dc]]]] [formerly 4fdbc06b5f9a4f4f97d4b13e404d884ffaadcb23 [formerly 0591290083aef69d5c061b129466b580f6fcb129] [formerly a25607bf66073ea184f0cbc344fc22525d24d3df [formerly ac0ad7b0f67a223bdcbf7c606150d3c217f72c84]] [formerly 9df70cc62c951166feef1b9eff807a87ff809655 [formerly 9ffa9c6b03a14109a0ed9a1359049a5de84b2935] [formerly b87b28da56cd59e07298c4faa1c354350fbbf3ed [formerly 4093ccb82068820c22195fa367685aa12df79aae]]] [formerly b5fddc3a9c22c26b5b14567b00305cc84a6958c3 [formerly 06f2fc2d6281a6503b6c71a925d7993f540d1832] [formerly 47f566541d9dd2bb4e3ae5a8a0de26342a8247b5 [formerly 1c84492c99316828b7a401376e8e6a89bc49460d]] [formerly 95f09a5b5ee96141e3ecb0dcfbb14a93eb3f5e8e [formerly 29505b45f836ce35cd8f39b0b8ad103581ed819b] [formerly 5e4c54897641e7bd6750d0469f49ca5d50f94878 [formerly 3133a7b523d012f3e1ad7077c8bf697fd20a1e7c]]]]]
Former-commit-id: 87e1fe21a6761e41c791c01e2fb357dc83e4a2b8 [formerly 55f099fcaf98ea9251ce32bdf1cba1dfff6c7e47] [formerly 5946e9f4e157bdc8b3dcfdbca63f86fc76be38ec [formerly 19600c0141fb068851d1960ff397346ed3032e4c]] [formerly a6f49da925fd0db93bab1237a8c24a0255459e3e [formerly 2733fd0e7f04bbe787ed39f00a39cb66a821a268] [formerly 242cfe60ac050e67d8fe88b236a335369d7d3df3 [formerly 614c16e71d34cb881aae19eed452573c93b38598]]] [formerly f3b95049fd5290c12332ae8390b5093a35e3c5fd [formerly ab783ef4c58a641b25dc70aff7b58caea49a5ef1] [formerly 3f14e354187a73cad2df0db36c0471fb3939692a [formerly 5c5e0767d34e25cbfa6e1c4ddd399646a45ae20e]] [formerly f4840cad44c7630b7b2d27e22cfad90c966905c0 [formerly 5cc7e724b065199b86cc8f225a2deef3e49720c1] [formerly 4b45a95907550f66b3e9c091a914ae6beba4d063 [formerly 6674ea33456579a283f750146928a7e71d21fa1f]]]]
Former-commit-id: df6551cd603fe2fc7bce50004d925028df48ea98 [formerly ba2839f9adbb67479e773a11147c4db7afb067e4] [formerly 1d3f7696a91f35ded3655e50dee0d0e52667f6dc [formerly a9c98256347b222e080c4abbf3a3723543f72048]] [formerly 8a80feb691bd196a541847973a45c5bb85ef5a7f [formerly 12d32092b60ca6f606498fbc51a4dc66ecb2e504] [formerly 98f84ab3920d7fed72e4d346b9ec165752743be4 [formerly d6ef32861f5e0372cc1c4be643ae31a7d543fd20]]]
Former-commit-id: 14a770c2a1bf03fd92fea370f209105a0be652d5 [formerly e5f25972e0d55b893cc7b26112a96175d7dc4b21] [formerly ac38f3813df5bd273f21503a47d61aa944705c03 [formerly efeaa93cd48ed626827bdf680244e506085ad263]]
Former-commit-id: beff2b5eb0db7eb936dac07ac69c1551365adaff [formerly c1bc4f7a1290065c22741faa5c527b4d8afe4f99]
Former-commit-id: f14a2b2b554f1def6478af244ac74b2b451062b8

+++ src/frame_buffer.h
@@ -14,14 +14,14 @@ class FrameBuffer {
       : width_(width),
         height_(height),
         data_(std::make_unique<unsigned char[]>(size())),
-        z_buffer_(std::make_unique<int[]>(width * height)) {
+        z_buffer_(std::make_unique<float[]>(width * height)) {
     ResetZBuffer();
   }
 
   ~FrameBuffer() = default;
 
-  void Set(int x, int y, Color const &color);
-  void Set(int x, int y, int z, Color const &color);
+  void Set(float x, float y, Color const &color);
+  void Set(float x, float y, float z, Color const &color);
   void Clear();
 
   int width() { return width_; }
@@ -37,7 +37,7 @@ class FrameBuffer {
   int width_{};
   int height_{};
   std::unique_ptr<unsigned char[]> data_{};
-  std::unique_ptr<int[]> z_buffer_{};
+  std::unique_ptr<float[]> z_buffer_{};
 };
 
 }  // namespace sren
